test: rename testMocks.ts to mocks.ts, move getMockPage into mocks.ts
Per @OrKoN review:
- Rename tests/testMocks.ts to tests/mocks.ts (already in tests folder)
- Move getMockPage() from tests/utils.ts into tests/mocks.ts
- tests/utils.ts re-exports getMockPage from mocks.ts for backwards compat
- Update emulation.test.ts and McpPage.test.ts imports accordingly
